Add more fdw features (#2)
* fdw test private stat, it works

* fix compile error. use faker to create random time.

* fix compile error. use faker to create random time.

* add std string for faker

* fdw options for total rows

* add random uuid and inet support

* change default guc

+++ examples/test.rs
@@ -1,15 +1,13 @@
-use std::hash::{Hash, Hasher};
 use std::collections::hash_map::DefaultHasher;
+use std::hash::{Hash, Hasher};
 
-
-
-fn my_hash<T>(obj: T) -> u128
-    where
-        T: Hash,
+fn my_hash<T>(obj: T) -> u64
+where
+    T: Hash,
 {
     let mut hasher = DefaultHasher::new();
     obj.hash(&mut hasher);
-    hasher.finish() as u128
+    hasher.finish()
 }
 
 fn main() {
